Create session context to update conditional user view on first render
